app/pyproject.toml
[tool.poetry]
name = "itmo-docker-task-5"
version = "0.1.0"
description = ""
authors = ["misha <mishaivanov99@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.85.0"
uvicorn = "^0.18.3"
SQLAlchemy = "^1.4.44"
aiomysql = "^0.1.1"
python-dotenv = "^0.21.0"
cryptography = "^38.0.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
